basic command line tool working
