--- src/writeDataToSVG.ts
@@ -1,8 +1,0 @@
-
-
-
-export function writeDataToSVGFile(data: string) {
-  const fileName = `./generated/translated_fr_${+new Date}.svg`
-  Bun.write(fileName, data)
-  return fileName
-}
